backend/requirements.txt
flask
flask-cors
Flask-Login
mysql-connector-python